+++ src/utils/componentCategories.ts
@@ -121,6 +121,7 @@ const categories = {
   navigation: {
     title: "Components/Navigation",
     components: [
+      "AlphabetFilter",
       "Breadcrumbs",
       "Link",
       "Menu",