+++ CHANGELOG.md
@@ -11,10 +11,13 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 ### Adds
 
 - Adds a variant for `NYCDOE` to the `Logo` component.
+- Adds the `TagSet` component.
+- Adds `AlphabetFilter` component.
 
 ### Updates
 
 - Updates the `hamburger` SVG for the `Icon` component to allow for color assignments.
+- Updates `ComponentWrapper` prop `descriptionText` type to `string | JSX.Element`.
 
 ## 1.1.2 (September 29, 2022)
 
@@ -92,7 +95,6 @@ This release is exactly the same as the `1.1.0-rc5` release candidate which was 
 - Adds @chakra-ui/focus-lock packages to dependencies and peerDependencies.
 - Adds the `useCloseDropDown` hook to manage closing dropdown-like components through clicks outside of it and by pressing the escape key.
 - Adds the `react-ga` package to send Google Analytics events in the `Header` component.
-- Adds the `TagSet` component.
 
 ### Updates